Created separate test file for Utils, moved JWT specific test to UtilsTest.php
